fix: remove more compiler warnigns

+++ src/openssl.cpp
@@ -73,7 +73,7 @@ bool AES_CCM_Decrypt(COSE_Enveloped *pcose,
 	EVP_CIPHER_CTX *ctx;
 	int cbOut;
 	byte *rgbOut = NULL;
-	int NSize = 15 - (LSize / 8);
+	size_t NSize = 15 - (LSize / 8);
 	int outl = 0;
 	byte rgbIV[15] = {0};
 	const cn_cbor *pIV = NULL;
@@ -175,7 +175,7 @@ bool AES_CCM_Encrypt(COSE_Enveloped *pcose,
 	EVP_CIPHER_CTX *ctx;
 	int cbOut;
 	byte *rgbOut = NULL;
-	int NSize = 15 - (LSize / 8);
+	size_t NSize = 15 - (LSize / 8);
 	int outl = 0;
 	const cn_cbor *cbor_iv = NULL;
 	cn_cbor *cbor_iv_t = NULL;
@@ -1060,7 +1060,7 @@ bool HMAC_Validate(COSE_MacMessage *pcose,
 	cn_cbor *cn = _COSE_arrayget_int(&pcose->m_message, INDEX_MAC_TAG);
 	CHECK_CONDITION(cn != NULL, COSE_ERR_CBOR);
 
-	if (cn->length > (int)cbOut) {
+	if (cn->length > cbOut) {
 		return false;
 	}
 	for (unsigned int i = 0; i < (unsigned int)TSize / 8; i++) {
@@ -1135,7 +1135,7 @@ EC_KEY *ECKey_From(COSE_KEY *pKey, int *cbGroup, cose_errback *perr)
 	p = cn_cbor_mapget_int(pKey->m_cborKey, COSE_Key_EC_X);
 	CHECK_CONDITION(
 		(p != NULL) && (p->type == CN_CBOR_BYTES), COSE_ERR_INVALID_PARAMETER);
-	CHECK_CONDITION(p->length == *cbGroup, COSE_ERR_INVALID_PARAMETER);
+	CHECK_CONDITION(p->length == (size_t)*cbGroup, COSE_ERR_INVALID_PARAMETER);
 	memcpy(rgbKey + 1, p->v.str, p->length);
 
 	p = cn_cbor_mapget_int(pKey->m_cborKey, COSE_Key_EC_Y);
@@ -1143,7 +1143,7 @@ EC_KEY *ECKey_From(COSE_KEY *pKey, int *cbGroup, cose_errback *perr)
 	if (p->type == CN_CBOR_BYTES) {
 		rgbKey[0] = POINT_CONVERSION_UNCOMPRESSED;
 		cbKey = (*cbGroup * 2) + 1;
-		CHECK_CONDITION(p->length == *cbGroup, COSE_ERR_INVALID_PARAMETER);
+		CHECK_CONDITION(p->length == (size_t)*cbGroup, COSE_ERR_INVALID_PARAMETER);
 		memcpy(rgbKey + p->length + 1, p->v.str, p->length);
 	}
 	else if (p->type == CN_CBOR_TRUE) {
